Moved create_z_from_label to utils.py
