+++ src/class_cvae/train_split_decoders.py
@@ -15,7 +15,7 @@ from PIL import Image
 from models import Encoder, Decoder, Classifier, ImageClassifier, SimpleEncoder, HandCraftedMNISTDecoder
 from logger import Logger
 from lpips.lpips import LPIPS
-from utils import init_weights, get_hardcode_mnist_latent_map
+from utils import init_weights, get_hardcode_mnist_latent_map, create_z_from_label
 
 """
 Goal: Train a variational autoencoder with a classification head on the latent space.
@@ -154,15 +154,6 @@ def get_args():
     parser.add_argument('--num_features', type=int, default=20)
     parser.add_argument('--num_class_features', type=int, default=7)
     return parser.parse_args()
-
-def create_z_from_label(lbls):
-    z_map = get_hardcode_mnist_latent_map()
-
-    z = z_map[lbls[0].item()]
-    for i in range(1, len(lbls)):
-        z = np.concatenate((z, z_map[lbls[i].item()]), axis=0)
-
-    return torch.tensor(z).cuda()
 
 if __name__ == "__main__":
     args = get_args()
